@@ -3,7 +3,7 @@
   "productName": "Podman Desktop",
   "repository": "https://github.com/podman-desktop/podman-desktop",
   "homepage": "https://www.podman-desktop.io",
-  "version": "1.22.0-next",
+  "version": "1.23.0-next",
   "license": "apache-2.0",
   "type": "module",
   "private": true,
@@ -118,8 +118,8 @@
     "*.{css,json}": "biome format --write"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.2.4",
-    "@commitlint/cli": "^20.0.0",
+    "@biomejs/biome": "^2.2.5",
+    "@commitlint/cli": "^20.1.0",
     "@commitlint/config-conventional": "^20.0.0",
     "@electron/fuses": "^2.0.0",
     "@eslint/compat": "^1.4.0",
@@ -142,13 +142,13 @@
     "@types/tar": "^6.1.13",
     "@types/tar-fs": "^2.0.4",
     "@types/validator": "^13.15.3",
-    "@typescript-eslint/eslint-plugin": "^8.44.1",
-    "@typescript-eslint/parser": "^8.44.1",
+    "@typescript-eslint/eslint-plugin": "^8.45.0",
+    "@typescript-eslint/parser": "^8.45.0",
     "@vitest/coverage-v8": "^3.2.4",
-    "@vitest/eslint-plugin": "^1.3.13",
-    "cross-env": "10.0.0",
+    "@vitest/eslint-plugin": "^1.3.15",
+    "cross-env": "10.1.0",
     "dts-for-context-bridge": "0.7.1",
-    "electron": "38.1.2",
+    "electron": "38.2.0",
     "electron-builder": "26.0.12",
     "electron-builder-notarize": "^1.5.2",
     "eslint": "^9.36.0",
@@ -166,17 +166,17 @@
     "get-tsconfig": "^4.10.1",
     "globals": "^16.4.0",
     "husky": "^9.1.7",
-    "inversify": "^7.10.1",
+    "inversify": "^7.10.2",
     "js-yaml": "^4.1.0",
-    "lint-staged": "^16.2.1",
+    "lint-staged": "^16.2.3",
     "minimist": "^1.2.8",
     "msw": "^2.11.2",
     "prettier": "^3.6.2",
     "proxy": "^2.2.0",
-    "typescript": "5.9.2",
-    "typescript-eslint": "^8.44.1",
+    "typescript": "5.9.3",
+    "typescript-eslint": "^8.45.0",
     "validator": "^13.15.15",
-    "vite": "^7.1.7",
+    "vite": "^7.1.9",
     "vitest": "^3.2.4",
     "xvfb-maybe": "^0.2.1"
   },
@@ -192,7 +192,7 @@
     "chokidar": "^3.6.0",
     "compare-versions": "^6.1.1",
     "date.js": "^0.3.3",
-    "dockerode": "^4.0.8",
+    "dockerode": "^4.0.9",
     "electron-context-menu": "4.1.1",
     "electron-updater": "6.6.8",
     "electron-util": "^1.0.0",
