@@ -24,8 +24,8 @@
   "devDependencies": {
     "@playwright/test": "1.55.1",
     "@types/node": "^22",
-    "electron": "38.1.2",
-    "typescript": "^5.9.2",
+    "electron": "38.2.0",
+    "typescript": "^5.9.3",
     "vitest": "^3.2.4"
   }
 }
