@@ -29,6 +29,8 @@ import NumberInput from './inputs/NumberInput.svelte';
 import SearchInput from './inputs/SearchInput.svelte';
 import DetailsPage from './layouts/DetailsPage.svelte';
 import FormPage from './layouts/FormPage.svelte';
+import type { ListOrganizerItem } from './layouts/ListOrganizer';
+import ListOrganizer from './layouts/ListOrganizer.svelte';
 import NavPage from './layouts/NavPage.svelte';
 import Page from './layouts/Page.svelte';
 import Link from './link/Link.svelte';
@@ -47,7 +49,7 @@ import Table from './table/Table.svelte';
 import Tooltip from './tooltip/Tooltip.svelte';
 import { isFontAwesomeIcon } from './utils/icon-utils';
 
-export type { ButtonType };
+export type { ButtonType, ListOrganizerItem };
 export {
   Button,
   Carousel,
@@ -64,6 +66,7 @@ export {
   Input,
   LinearProgress,
   Link,
+  ListOrganizer,
   Modal,
   NavPage,
   NumberInput,
